==> 1.Preprocessing Raw Data.ipynb <==
# 1-Preprocess Candlesticks Data

### Import Data

import pandas as pd
import numpy as np
import os

# Folder path containing CSV files
folder_path = r"C:\Users\Mcc\Desktop\Data\Historical Data\Candlesticks\BTCUSDT-2024-04-01 - 2024-04-17\5M"
column_names = ["open_time", "open", "high", "low", "close", "volume", "close_time", "quote_volume", 'count', 'taker_buy_volume', 'taker_buy_quote_volume', 'ignore']
# Create an empty list to store DataFrames
dfs = []

# Iterate through each file in the folder
for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path
        #df = pd.read_csv(file_path, header=None, names=column_names)  # Read the CSV file into a DataFrame
        df = pd.read_csv(file_path) 
        dfs.append(df)  # Append the DataFrame to the list

# Concatenate all DataFrames in the list along the rows (axis=0) to merge them
df_candle =  pd.concat(dfs, axis=0, ignore_index=True)
df_candle.to_csv('BTCUSDT-5m-2020-01-2024-03.csv')

df_candle

df_candle.drop(columns=['Unnamed: 0'], inplace=True)

df_candle['open_time'] = pd.to_datetime(df_candle['open_time'], unit='ms')
df_candle.set_index('open_time', inplace=True)

df_candle.drop(columns=['close_time'], inplace=True)
#df_candle['close_time'] = pd.to_datetime(df_candle['close_time'], unit='ms')

df_candle.drop(columns=['ignore'], inplace=True)

df_candle

df_candle.to_csv('BTCUSDT-2024-04-01 - 2024-04-17 - 5M.csv')

# 2-Preprocessing Agg Data

### Import & Turn the Fractions into 1 & 5 minutes Candles of True & False Buyer Makers + Vdiff

import os
import pandas as pd
import numpy as np

folder_path = r"C:\Users\Mcc\Desktop\Data\Historical Data\Agg Volumes\BTCUSDT-aggTrades-2024-04-01 - 2024-04-17\COIN-M"
column_names = ["agg_trade_id", "price", "quantity", "first_trade_id", "last_trade_id", "transact_time", "is_buyer_maker"]

def remove_milliseconds(datetime_str):
    return datetime_str.strftime('%Y-%m-%d %H:%M:%S')

def generate_minute_volumes(df_agg, is_buyer_maker):
    df_filtered = df_agg.loc[df_agg['is_buyer_maker'] == is_buyer_maker]
    df_minute_volumes = df_filtered.resample('5T')['volume'].sum().fillna(0)
    #print(df_minute_volumes)
    return df_minute_volumes

for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path

        #df_agg = pd.read_csv(file_path, header=None, names=column_names)
        df_agg = pd.read_csv(file_path)
        df_agg['transact_time'] = pd.to_datetime(df_agg['transact_time'], unit='ms')
        df_agg.set_index('transact_time', inplace=True)
        df_agg['volume'] = df_agg['price'] * df_agg['quantity']

        start_time_str = df_agg.index[0]
        end_time_str = df_agg.index[-1]

        df_maker_true = generate_minute_volumes(df_agg, True)
        df_maker_false = generate_minute_volumes(df_agg, False)

        merged_df = pd.merge(df_maker_true, df_maker_false, left_index=True, right_index=True, how='outer', suffixes=('_true', '_false'))
        merged_df['vdiff(false-true)'] = merged_df['volume_false'] - merged_df['volume_true']
        #merged_df.index = merged_df.index.floor('s')

        # Extract filename without extension
        filename = os.path.splitext(file)[0]
        new_filename = f"{filename}-COINM-5M.csv"  # New CSV filename
        merged_df.to_csv(new_filename)
        current_time = pd.Timestamp.now().strftime("%Y-%m-%d %H:%M:%S")
        print(f"Finished & Saved - {new_filename} - at {current_time}")


### Concat 1 & 5 Minutes Candles into 1 Dataframe & CSV File

import pandas as pd
import numpy as np
import os

folder_path = r"C:\Users\Mcc\Desktop\Project\Data Preprocessing for ML\New folder"
# Create an empty list to store DataFrames
dfs = []

# Iterate through each file in the folder
for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path
        df = pd.read_csv(file_path) 
        dfs.append(df)  # Append the DataFrame to the list

# Concatenate all DataFrames in the list along the rows (axis=0) to merge them
df_agg =  pd.concat(dfs, axis=0, ignore_index=True)
df_agg.to_csv('BTCUSDT-aggTrades-2024-04-01-2024-04-17 -COINM- Resampled 5M.csv')

df_agg

# 3-Preprocessing Funding Rate Data

### Import Data

import pandas as pd
import numpy as np
import os

# Folder path containing CSV files
folder_path = r"C:\Users\Mcc\Desktop\Data\Historical Data\Fundaing Rate"

# Create an empty list to store DataFrames
dfs = []

# Iterate through each file in the folder
for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path
        df = pd.read_csv(file_path)  # Read the CSV file into a DataFrame
        dfs.append(df)  # Append the DataFrame to the list

# Concatenate all DataFrames in the list along the rows (axis=0) to merge them
df_fund =  pd.concat(dfs, axis=0, ignore_index=True)
df_fund.to_csv('BTCUSDT-fundingRate-2020-01-2024-03.csv')

df_fund

df_fund['calc_time'] = pd.to_datetime(df_fund['calc_time'], unit='ms')
df_fund.set_index('calc_time', inplace=True)

df_fund.to_csv('BTCUSDT-fundingRate-2020-01-2024-03.csv')

df_fund

### Resample to 1 & 5 Minutes

import pandas as pd

# Assuming df is your DataFrame with the funding rates
# Assuming calc_time is the index of your DataFrame

# Resample to 1-minute frequency and forward fill missing values
df_resampled = df_fund.resample('1T').ffill()

# Generate the complete date range from the start to the end of your desired period
start_date = '2020-01-01 00:00:00.000'
end_date = '2024-03-31 16:00:00.001'
date_range = pd.date_range(start=start_date, end=end_date, freq='5T')

# Reindex the DataFrame with the complete date range
df_complete = df_resampled.reindex(date_range)

# Fill any remaining missing values with the last available value
df_complete = df_complete.ffill()

# If needed, reset index to have 'calc_time' as a column instead of index
df_complete.reset_index(inplace=True)

# Rename the columns if needed
# df_complete = df_complete.rename(columns={'index': 'calc_time'})

# Now df_complete contains funding rates at 1-minute frequency from 2020-01-01 to 2024-03-31 16:00:00.001


df_complete

df_complete.to_csv('BTCUSDT-fundingRate-2020-01-2024-03- Resampled 5M.csv')

# 4-Preprocessing Trading Metrics

import pandas as pd
import numpy as np
import os

# Folder path containing CSV files
folder_path = r"C:\Users\Mcc\Desktop\Data\Historical Data\Trading Metrics (Open Interest,Long Short Ratio)"
#column_names = ["open_time", "open", "high", "low", "close", "volume", "close_time", "quote_volume", 'count', 'taker_buy_volume', 'taker_buy_quote_volume', 'ignore']
# Create an empty list to store DataFrames
dfs = []

# Iterate through each file in the folder
for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path
        #df = pd.read_csv(file_path, header=None, names=column_names)  # Read the CSV file into a DataFrame
        df = pd.read_csv(file_path) 
        dfs.append(df)  # Append the DataFrame to the list

# Concatenate all DataFrames in the list along the rows (axis=0) to merge them
df_metrics =  pd.concat(dfs, axis=0, ignore_index=True)
df_metrics.to_csv('BTCUSDT-metrics-2023-12-01-2024-03-31.csv')

df_metrics

# 5-Preprocessing Liquidation

import pandas as pd
import numpy as np
import os

# Folder path containing CSV files
folder_path = r"C:\Users\Mcc\Desktop\Data\Historical Data\Liquidation\BTCUSD_COINM-liquidationSnapshot-2024-04-01 - 2024-04-17"
#column_names = ["open_time", "open", "high", "low", "close", "volume", "close_time", "quote_volume", 'count', 'taker_buy_volume', 'taker_buy_quote_volume', 'ignore']
# Create an empty list to store DataFrames
dfs = []

# Iterate through each file in the folder
for file in os.listdir(folder_path):
    if file.endswith('.csv'):  # Check if the file is a CSV file
        file_path = os.path.join(folder_path, file)  # Get the full file path
        #df = pd.read_csv(file_path, header=None, names=column_names)  # Read the CSV file into a DataFrame
        df = pd.read_csv(file_path) 
        dfs.append(df)  # Append the DataFrame to the list

# Concatenate all DataFrames in the list along the rows (axis=0) to merge them
df_liquid =  pd.concat(dfs, axis=0, ignore_index=True)
df_liquid['time'] = pd.to_datetime(df_liquid['time'], unit='ms')
df_liquid.set_index('time', inplace=True)
df_renamed_index = df.rename_axis('date')
df_liquid['original_volume'] = df_liquid['price'] * df_liquid['original_quantity']
df_liquid['last_fill_volume'] = df_liquid['price'] * df_liquid['last_fill_quantity']
df_liquid['accumulated_fill_volume'] = df_liquid['price'] * df_liquid['accumulated_fill_quantity']
df_liquid.to_csv('BTCUSDT-liquidationSnapshot-2023-12-01-2024-03-31-RAW.csv')

df_liquid

df_1m_liq = df_liquid.resample('5T').sum().fillna(0)
df_1m_liq = df_1m_liq.drop(columns=['order_status','time_in_force','order_type','time_in_force','price','average_price',"original_quantity",'last_fill_quantity','accumulated_fill_quantity'])

# Define a function to check if the value contains 'SELL'
def process_side(value):
    if isinstance(value, str) and 'SELL' in value:
        return 'SELL'
    elif isinstance(value, str) and 'BUY' in value:
        return 'BUY'
    else:
        return value

# Apply the function to the 'side' column
df_1m_liq['side'] = df_1m_liq['side'].apply(process_side)


df_1m_liq.to_csv('BTCUSDT-liquidationSnapshot-2024-04-01-2024-04-17-Resampled 5M.csv')

df_1m_liq
