==> 2.Combine Multiple Datasets into One.ipynb <==
import pandas as pd
import numpy as np

df_canle = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\Candlesticks\Completed\BTCUSDT-2024-04-01 - 2024-04-17 - 5M.csv")
df_agg = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\Agg Volumes\Completed\BTCUSDT-aggTrades-2024-04-01-2024-04-17 -COINM- Resampled 5M.csv")
# df_fund = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-fundingRate-2020-01-2024-03- Resampled 1M.csv")
df_liquid = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\Liquidation\Completed\BTCUSDT-liquidationSnapshot-2024-04-01-2024-04-17-Resampled 5M.csv")

df_canle = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-5m-2020-01-2024-03.csv")
df_agg = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-aggTrades-2020-01-2024-03 - Resampled 5M.csv")
# df_fund = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-fundingRate-2020-01-2024-03- Resampled 5M.csv")
df_liquid = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-liquidationSnapshot-2023-12-01-2024-03-31-Resampled 5M.csv")
# df_metrics = pd.read_csv(r"C:\Users\Mcc\Desktop\Data\Historical Data\.All in One\BTCUSDT-metrics-2023-12-01-2024-03-31-Resample 5M.csv")

df_canle

df_agg

df_liquid

# df_canle = df_canle.drop(columns=["ignore"])
df_agg = df_agg.drop(columns=["Unnamed: 0"])
# df_fund = df_fund.drop(columns=["Unnamed: 0",'funding_interval_hours'])
# df_metrics = df_metrics.drop(columns=["Unnamed: 0"])

df_canle = df_canle.rename(columns={"open_time": "date_time"})
df_agg = df_agg.rename(columns={"transact_time": "date_time"})
# df_fund = df_fund.rename(columns={"index": "date_time"})
df_liquid = df_liquid.rename(columns={"time": "date_time"})
#df_metrics = df_metrics.rename(columns={"create_time": "date_time"})

df_canle["log"] = np.log(df_canle["close"])
df_canle["returns"] = df_canle["low"].pct_change()
df_canle["range"] = (df_canle["high"] / df_canle["low"]) - 1

all_in_one_p1 = pd.merge(df_canle,df_agg, on="date_time", how="inner")
all_in_one_p2 = pd.merge(all_in_one_p1, df_liquid, on="date_time", how="inner")
# all_in_one_p3 = pd.merge(all_in_one_p2,df_metrics, on="date_time", how="inner")
# all_in_one_p4 = pd.merge(all_in_one_p1, all_in_one_p3, on="date_time", how="inner")

all_in_one_p2

all_in_one_p2.dropna(inplace=True)
all_in_one_p2.set_index("date_time", inplace=True)

all_in_one_p2.to_csv('BTC-Dataset-COINM-2024-04-01 - 2024-04-17 - 5M.csv')
