==> 3.Adding Price Based Indicators.ipynb <==
# IMPORTS

import pandas as pd
import numpy as np

df= pd.read_csv(r"BTC-Dataset-USDM-2024-04-01 - 2024-04-17 - 5M.csv")

df

# Define Price Based Indicators

# Function to calculate Exponential Moving Averages (EMAs)
def calculate_ema(data, periods=[12, 26, 50, 100, 200, 500]):
    for period in periods:
        ema_name = f'EMA_{period}'
        data[ema_name] = data['close'].ewm(span=period, adjust=False).mean()
    return data

# Function to calculate Relative Strength Index (RSI)
def calculate_rsi(data, window=14):
    delta = data['close'].diff()
    gain = (delta.where(delta > 0, 0)).rolling(window=window).mean()
    loss = (-delta.where(delta < 0, 0)).rolling(window=window).mean()
    rs = gain / loss
    rsi = 100 - (100 / (1 + rs))
    return rsi

# Function to calculate Stochastic Oscillator
def calculate_stochastic(data, k_period=14, d_period=3):
    lowest_low = data['low'].rolling(window=k_period).min()
    highest_high = data['high'].rolling(window=k_period).max()
    data['stochastic_%K'] = ((data['close'] - lowest_low) / (highest_high - lowest_low)) * 100
    data['stochastic_%D'] = data['stochastic_%K'].rolling(window=d_period).mean()
    return data

# Function to calculate Moving Average Convergence Divergence (MACD)
def calculate_macd(data, short_window=12, long_window=26, signal_window=9):
    short_ema = data['close'].ewm(span=short_window, adjust=False).mean()
    long_ema = data['close'].ewm(span=long_window, adjust=False).mean()
    data['MACD'] = short_ema - long_ema
    data['MACD_signal'] = data['MACD'].ewm(span=signal_window, adjust=False).mean()
    return data

# Function to calculate Bollinger Bands
def calculate_bollinger_bands(data, window=20, num_std=2):
    rolling_mean = data['close'].rolling(window=window).mean()
    rolling_std = data['close'].rolling(window=window).std()
    data['Bollinger_Band_Upper'] = rolling_mean + (rolling_std * num_std)
    data['Bollinger_Band_Lower'] = rolling_mean - (rolling_std * num_std)
    return data

# Add The Indicators to Dataframe

# Calculate RSI
df['RSI'] = calculate_rsi(df)

# Calculate Stochastic Oscillator
df = calculate_stochastic(df)

# Calculate EMAs
df = calculate_ema(df)

# Calculate MACD
df = calculate_macd(df)

# Calculate Bollinger Bands
df = calculate_bollinger_bands(df)


# Display the updated DataFrame
df.tail(5)

# Clean The Dataframe & Remove Unneccassary Columns

df['close_pct'] = df['close'].pct_change() * 100
df = df.drop(columns=['Unnamed: 0', 'side', 'high','low','close','open','taker_buy_volume','last_fill_volume', 'accumulated_fill_volume','volume','taker_buy_quote_volume'])
df.dropna(inplace=True)
df.set_index('date_time', inplace=True)

df
